Create basic game
I created a class Game and upon initialization the following instance
variables are set
@solution takes one word between 5 to 12 letters from a .txt file and
turns it into an array of single letters
@guessed is an array indicating the length of the word using _ and
storing the correct letters when guesses at the right place
@ended is used to check whether to @solution has been found
@guess will store the letter guessed by the player
@alphabet is an array of all letters to check whether @guess is valid
@wrong_guesses is used to track the number of wrong tries available

I wrote a method to display the @guessed array so the player can see all
the correct guesses and the length of the @solution
Then the player is asked to make a @guess and the guessed letter is
compared to each of the @solution letters. If there is a match, the same
index of @guessed is set to @guess.
If the letter is not included, the number of @wrong_guesses is reduced
by one.
At last the @guess is removed from @alphabet to make prevent duplicate
guesses

If @guessed == @solution, the player wins and a message is displayed.
If the player is out of @wrong_guesses, the player lost and a different
message gets displayed