Add save and load option
I added a method to the class that uses JSON.dump to save all relevant
instance variables inside an array to a file as a string.
Then I wrote another method outside the class that uses JSON.load to
open the file and turn it back into an array.

Upon starting the game, the player is asked whether to start a new game
or load the saved game.
If the player decides to load the game, an instance of the game class is
created which uses all the instance variables from the saved game